pyproject.toml
[project]
name = "crypto-prediction-bot"
version = "1.0.0"
description = "Transformer-based cryptocurrency prediction bot with backtesting"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.4.0",
    "yfinance>=0.2.0",
    "requests>=2.31.0",
    "python-binance>=1.0.0",
    "backtrader>=1.9.78",
    "matplotlib>=3.7.0",
    "plotly>=5.14.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=src --cov-report=html --cov-report=term"
minversion = "7.0"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
fail_under = 10
# Note: Target is 90% for core logic modules (bin_calculator, data_processor, etc.)
# Full codebase coverage tracked separately

